requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5

# Database
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Task queue
celery[redis]==5.4.0
redis==5.2.1

# AI
openai==1.58.1

# Telegram
python-telegram-bot==21.9

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
